+++ CHANGELOG.md
@@ -13,7 +13,9 @@ Notable changes of tdmclient. Release versions refer to [https://pypi.org/projec
     - Transpiler exception for unsupported augmented assignments `/=`, `**=` and `@=`
     - Errors on undefined variables or invalid use of list variables or indexing in augmented assignments
 - In function `ClientAsync.wait_for_node`, optional argument `timeout`
-- In `tdmclient.tools.server`, option `--port`
+- In `tdmclient.tools.server`:
+    - option `--port`
+    - support for WebSocket
 
 ### Fixed